Move solutions-related data structures and functions into `solutions` module
