+++ crates/sc-consensus-subspace-rpc/src/lib.rs
@@ -57,7 +57,8 @@ use subspace_core_primitives::hashes::Blake3Hash;
 use subspace_core_primitives::objects::GlobalObjectMapping;
 use subspace_core_primitives::pieces::{Piece, PieceIndex};
 use subspace_core_primitives::segments::{HistorySize, SegmentHeader, SegmentIndex};
-use subspace_core_primitives::{BlockHash, PublicKey, SlotNumber, Solution};
+use subspace_core_primitives::solutions::Solution;
+use subspace_core_primitives::{BlockHash, PublicKey, SlotNumber};
 use subspace_erasure_coding::ErasureCoding;
 use subspace_farmer_components::FarmerProtocolInfo;
 use subspace_kzg::Kzg;
