+++ crates/subspace-core-primitives/src/sectors/tests.rs
@@ -1,5 +1,5 @@
+use crate::pieces::Record;
 use crate::sectors::SBucket;
-use crate::Record;
 
 // Statically validate that we can store all possible s-buckets in SBucket data structure
 #[test]
