+++ crates/subspace-farmer-components/src/auditing.rs
@@ -12,7 +12,8 @@ use std::collections::HashSet;
 use std::io;
 use subspace_core_primitives::hashes::Blake3Hash;
 use subspace_core_primitives::sectors::{SBucket, SectorId, SectorIndex, SectorSlotChallenge};
-use subspace_core_primitives::{PublicKey, ScalarBytes, SolutionRange};
+use subspace_core_primitives::solutions::SolutionRange;
+use subspace_core_primitives::{PublicKey, ScalarBytes};
 use subspace_verification::is_within_solution_range;
 use thiserror::Error;
