+++ crates/subspace-farmer-components/src/proving.rs
@@ -17,7 +17,8 @@ use std::io;
 use subspace_core_primitives::pieces::{PieceOffset, Record};
 use subspace_core_primitives::pos::PosSeed;
 use subspace_core_primitives::sectors::{SBucket, SectorId};
-use subspace_core_primitives::{ChunkWitness, PublicKey, ScalarBytes, Solution, SolutionRange};
+use subspace_core_primitives::solutions::{ChunkWitness, Solution, SolutionRange};
+use subspace_core_primitives::{PublicKey, ScalarBytes};
 use subspace_erasure_coding::ErasureCoding;
 use subspace_kzg::Kzg;
 use subspace_proof_of_space::Table;
