+++ crates/subspace-farmer/src/bin/subspace-farmer/commands/benchmark.rs
@@ -10,7 +10,7 @@ use std::num::NonZeroUsize;
 use std::path::PathBuf;
 use subspace_core_primitives::hashes::Blake3Hash;
 use subspace_core_primitives::pieces::Record;
-use subspace_core_primitives::SolutionRange;
+use subspace_core_primitives::solutions::SolutionRange;
 use subspace_erasure_coding::ErasureCoding;
 use subspace_farmer::single_disk_farm::direct_io_file::DirectIoFile;
 use subspace_farmer::single_disk_farm::farming::rayon_files::RayonFiles;
