+++ crates/subspace-farmer/src/single_disk_farm/farming.rs
@@ -23,7 +23,8 @@ use subspace_core_primitives::pieces::Record;
 use subspace_core_primitives::pos::PosSeed;
 use subspace_core_primitives::sectors::SectorIndex;
 use subspace_core_primitives::segments::{HistorySize, SegmentIndex};
-use subspace_core_primitives::{PublicKey, Solution, SolutionRange};
+use subspace_core_primitives::solutions::{Solution, SolutionRange};
+use subspace_core_primitives::PublicKey;
 use subspace_erasure_coding::ErasureCoding;
 use subspace_farmer_components::auditing::{audit_plot_sync, AuditingError};
 use subspace_farmer_components::proving::{ProvableSolutions, ProvingError};
