+++ crates/subspace-verification/src/lib.rs
@@ -44,10 +44,9 @@ use subspace_core_primitives::sectors::SectorSlotChallenge;
 use subspace_core_primitives::segments::ArchivedHistorySegment;
 use subspace_core_primitives::segments::{HistorySize, SegmentCommitment};
 #[cfg(feature = "kzg")]
-use subspace_core_primitives::Solution;
-use subspace_core_primitives::{
-    BlockNumber, BlockWeight, PublicKey, RewardSignature, ScalarBytes, SlotNumber, SolutionRange,
-};
+use subspace_core_primitives::solutions::Solution;
+use subspace_core_primitives::solutions::{RewardSignature, SolutionRange};
+use subspace_core_primitives::{BlockNumber, BlockWeight, PublicKey, ScalarBytes, SlotNumber};
 #[cfg(feature = "kzg")]
 use subspace_kzg::{Commitment, Kzg, Scalar, Witness};
 #[cfg(feature = "kzg")]
@@ -139,7 +138,7 @@ fn calculate_solution_distance(
             .next()
             .expect("Solution range is smaller in size than global challenge; qed"),
     );
-    subspace_core_primitives::bidirectional_distance(
+    subspace_core_primitives::solutions::bidirectional_distance(
         &global_challenge_as_solution_range,
         &audit_chunk_as_solution_range,
     )
