+++ test/subspace-test-client/src/lib.rs
@@ -38,7 +38,8 @@ use subspace_core_primitives::objects::BlockObjectMapping;
 use subspace_core_primitives::pieces::Record;
 use subspace_core_primitives::pos::PosSeed;
 use subspace_core_primitives::segments::{HistorySize, SegmentIndex};
-use subspace_core_primitives::{PublicKey, RewardSignature, Solution, REWARD_SIGNING_CONTEXT};
+use subspace_core_primitives::solutions::{RewardSignature, Solution};
+use subspace_core_primitives::{PublicKey, REWARD_SIGNING_CONTEXT};
 use subspace_erasure_coding::ErasureCoding;
 use subspace_farmer_components::auditing::audit_sector_sync;
 use subspace_farmer_components::plotting::{
